vinoteca_ia/pyproject.toml
[project]
name = "vinoteca-ia"
version = "1.0.0"
description = "Agente sumiller con sistema multi-agente para vinoteca"
requires-python = ">=3.11"
dependencies = [
    "agno>=2.5.16",
    "pydantic>=2.7.0,<3.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "asyncpg>=0.30.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "httpx>=0.27.0",
    "redis[hiredis]>=5.0.0",
    "pgvector>=0.3.0",
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "motor>=3.4.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.4.0",
    "anyio>=4.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
